statusbar: Add back PanelView.isFullyExpanded()
* This is a small part of the original commit (below), which only adds back
some of the missing functionality:

Changed the overscroll and expanding behaviour.

Only the first selected element will be expanded, no subsequent children.
Afterwards, overscrolling is performed.
This improves overscroll consistency a lot and people don't accidentally
expand unwanted notifications, just the one they wanted to.
If the users primary intent is overscrolling (i.e if he drags on a card
which is already expanded), then we allow him to go to the quick settings.

Bug: 14487435
Bug: 15181651
Original-Change-Id: I978cc4e06ae85c2ca69e15a149cb85ac54b2ef35

Change-Id: Ifbd43af86a63ef8379fdb954cf19bfa36346a25e
Signed-off-by: Dan Cartier <nospamdan1@gmail.com>